Add recorded_at to example campaign_forms
Campaign forms need to be recorded (i.e. non-draft) in order to show up
on the unmatched consent responses page. The generator also has to be
changed to ensure that a predictable timestamp is generated, including
the time component which is new.

The example campaign generator spec already requires that the date the
current example campaign was generated be included in it. Instead of
adding the time to this to ensure we have consistent timestamps, I've
opted to freeze time at 12:00:00 on the current day in the campaign
generator.

Pros (versus other methods such as putting the full time into the spec):
- we have a realistic-looking timestamp that's in the last 14 days
- we don't have to specify the exact time in the spec which is more
  fiddly than just adding the date