Enable flow logs via copilot
Identified as missing in the pen test. This copilot option is documented
here:

https://aws.github.io/copilot-cli/docs/manifest/environment/#network-vpc-flowlogs

Further reading about flow logs:

https://docs.aws.amazon.com/vpc/latest/userguide/flow-logs.html

Seems possible this will require reprovisioning the environments as the
VPCs are created for each env. You can check if flow logs are enabled by
going to each VPC in the console and checking the "Flog logs" tab.