Add feature initializers to flipper initializer
The Flipper initializer feels like a more obvious place to define an
initializer for Flipper featuers, vs the seeds.rb. The problem is Rails
initializers can't assume the db is available. By defining a proc in the
inialiizer but only running it when we need to use the Flipper seems to
be the best compromise approach.