Remove default campaign from session create
To bootstrap the session creation journey we used a default campaign
when creating a session. Now that we have a form page to set the
campaign we don't want to use a default, but this has implications for
other parts of the session creation journey that relied on the campaign
existing.