Fix campaign loader
The `else` block never gets triggered as all of our example files
contain a `sessions` key. So the mapping of `sendConsentAt` ->
snake_cased versions didn't work before. The reason we didn't catch this
is because we relied on `in_progress=true` which takes the importer down
a different path that manually specifies the timestamps.