Use date fields for send_consent_at
This replaces the field that asks for a day before the session to one
that simply asks for a specific day/month/year.

One bug with this approach is that the `DateParamsValidator` comes
before the validations on the model, so if there are errors there, the
errors on the model don't get surfaced and re-appear once the date is
entered correctly. I'd like to revisit the `DateParamsValidator`
implementation at some point and hopefully get it working more similarly
to other validators.