Fix error link to terms in registration form
The exact setup of "I agree" checkboxes is somewhat incompatible with
the form builder. This is a hack to fix the error message link. Worth
noting that it's not a full fix, as it would be better if the focus is
on the first checkbox rather than the fieldset. However, trying that
leads to the label breaking, and is probably more involved.