Add testing controller and generate_campaign
This introduces a new approach to creating test campaigns. Instead of
relying on a JSON file which is updated for any changes to our data
model, and imported for each test, this creates a specialised testing
controller that can be used to create the necessary campaigns that are
used for a test. Perceived benefits are:

- no need to update all the code we currently use generate and import
  the JSON file ... that code is getting out of hand and difficult to
  maintain
- as a corollary to the above, we define our test data closer to the
  code that is testing it
- also; saner fixtures
- and ... by creating a new campaign for each test, we should be able to
  parallelise the Playwright tests, hopefully speeding up their running.
  This is as of yet unproven though