Increase cpu, memory and task count for staging
This is to test before going to production. We should have 2 instances
running in production anyway, and the cpu and memory are increased to
also accomadate running Good Job processes in async mode.

We should lower the cpu and memory, but leave the count at 2, once we've
successfully updated production.