Ensure example patients have a location
Our current data model ties patients to a location: the way to get
patients into the system is up uploading a cohort list, and that ties
the patient to a location. This change ensures that our example data
matches this model.

This will also be used as part of the policy scope for patients,
only allowing nurses to see the patients tied to a location that is
being used by one of the sessions they own (via their team).