Add ConsentForm#find_matching_patient
This attempts to match a consent form to a patient using the matching
criteria we've defined for the pilot. If any 3 of first_name, last_name,
date_of_birth, address_postcode are exact matches, then we can safely
assume that the ConsentForm belongs to that particular Patient record.

A few things are left to be done in follow-up work:

- Case insensitive matching
- Additional criteria (such as using common_name)
- Creating Consent records (combined with Consent#from_consent_form)