Use parent name from consent in triage email
The consent object gets the parent's name from the consent form or from
the nurse when they call the parent. This is more likely to be accurate
and up-to-date than what's on the patient record.

There's still the question of what to do when more than one parent has
submitted consent, how do we know we have the right parent?