Run good_job in external mode using supervisor
This attempts to fix the current issue with running `good_job` in async
mode. It switches to external mode and runs `bundle exec good_job start`
as a separate process in the Docker container.

Because Docker containers are designed to be single process, we need to
use `supervisor` to manage running the two processes.