Simplify last_action lookup in AppStatusBanner
We're going to make a change that will set the `patient_session` state
to `unable_to_vaccinate`, but without a `vaccination_record` attached.
This would break the status banner component, because `last_action`
reads the state and decides that the `vaccination_record` is the thing
to show in that state.

Instead of relying on `state`, we can rely on the presence of
`vaccination_record`.