Before the vaccination event, open the health questions if there are notes
This is especially important during vaccinations, when staff need to double check what notes the parents left.

If they know that health questions can be closed even with notes, they would have to manually open every patient's health questions to see if there are any notes there or not.

After vaccination, the parent's notes are less likely to be needed to don't need to be opened by default.