Add state concern to replace state machine
Now that we understand how states work better, and put all the states we
need into the various objects connected to PatientSessions, it seems
clear that the state machine isn't actually adding anything beyond
additional hurdles and a false sense of comfort.

The new approach in this concern is something we discussed in the
beginning; we "calculate" the PatientSession state by looking at the
state of the various connected objects: consent, triage and
vaccination_record. This was all being done in the state machine as it
stands, anyway, the only additional complexity is the order in which we
look at these objects. What's removed, really, are the transitions, and
all the code to define, manage and test them.