Ensure we include/preload patient_session assocs
Now that we calculate the patient_session state on-the-fly, we're more
reliant on having PatientSession associations. The request response
time, as calculated by Rails, basically doubled.

Always pulling in these associations reduces our request response time
by about 25% for the consent, triage and vaccinations pages. I'm not
completely clear why this wasn't already being caught out by the
"strict_loading" we have, but I think it may be because we were using
"consents.recorded" or "patient_sessions.group_by" which was triggering
an implicit includes. However, we're removing the use of the "recorded"
scope in a follow-up PR because it ironically ends up triggering
additionaly requests.