Seed health questions if consent given
470bf3098 added a callback to the ConsentForm model that removes health
questions when a user changes their response from given to refused.

However, because health answers were being initialised on the create
action in the controller, this meant that if a user then subsequently
changes their response from refused to given, the `health_answers` array
would be empty. So users would be unable to answer health questions.

This moves the initialisation logic from the controller to the model and
makes it consistent with the removal logic.