Patch nhsuk-password-input JS
The form builder has been updated to include new GOVUK Password Input
HTML which adds a button that can be used to toggle the input type from
password to plain text.

The button introduced a few issues. Due to flex layout, the extra bottom
margin was causing the input to stretch vertically, which is addressed
with some overrides.

The custom JS was trickier as the GOVUK JS relies on GOVUK specific
classes, and the form builder is replacing those with NHSUK ones that
don't actually mean anything. To address this, the stimulus controller
shims the classes.