Update tests for triage tabs changes
The triage tabs no longer behave as before (clicking on them reloads the
page, causing the alert to disappear, for example), and break the aria
roles assigned by nhsuk-frontend. The latter will be addressed later
when we re-implement the tab designs, for now the tests need to just
look for "links" instead of "tabs".