Fix back link on manage consents journey
When nurses amend an existing consent, the link sends them to the middle
of a consent wizard. `previous_wizard_path` links to the first step in
the journey, which causes the back button to go into a loop, because the
first step is the :agree one which the user is currently on.

This adds a special `back_link_path` helper that checks for this case
and works around it by instead sending a user to the session path for
their route. The code to do this is extracted from the :who step. This
way, every step is now a possible "first step" depending on how the
logic to handle the list of `form_steps` changes.