Update paths in components, views and controllers
For the most part this adds the section and tab parameters and updates
the names of the URL helpers used to match the new routes. In some
instances the session or patient params need to have keywords to work
with the newly helpers (because of the added section and tab params).

We also take the opportunity to clean up how paths are generated in a
few instances, since we can parameterise section and tab now.