Add PatientsController to replace PatientSessions
We've been rendering a patient from the patient_sessions controller for
some time now. In some ways this is wrong (we render out information for
a patient and a specific session), this can get confusing since
patient_sessions are their own entities in the db.

With the routes refactoring it makes sense to move to using the patients
controller to render patients, at least for now while our routes are
like this. The change to the routes will happen in a later commit.