Add example_campaigns factory and use from /reset
With the example campaign JSON gone, the /reset endpoint will no longer
be used to generate test data for autmoated tests. It will only be used
for manual testing. This will allow us to remove code for the generation
of the JSON, and so far appears to be faster. This approach has also
come with the extra benefits of not logging us out after running, and
has exposed some issues with our factories.