Stop creating extra records in patient_sessions
When patient_session isn't specified for the triage factory, it created
it's own, which was immediately overwritten by the association
assignment. The result is an extra team, campaign, vaccine, batch, session,
and location records for each patient_session created. In the example
campaign, this means 5 of each of those records being created instead of
just 1.

This change ensures only the necessary patient_session (and associated
records) are created for this trait.

In unit tests this results in a tests speedup as well.