Stop using after :create blocks in factories
A lot of the "after :create" blocks in our factories were being used to
create associations that are has_one/has_many: since associated model
has a belongs_to that relies on the factory model, we thought the best
way to ensure these are created was in the after :create.

Factory bot provides a way to do this without the after :create by using
"instance" for the factory model being instantiated. This means we can
override associations when creating new instances of models, preventing
us from creating extra unnecessary records.