Add session in the past to example campaign
Based on the in_progress factory trait, this adds a session in the
past and some patients in the various appropriate states.

Because of how the example campaign factory is structured, we have to
use multiple traits to get the different sessions when creating the
example campaign. This works so far, but we may have to change how this
works.