Disable turbo globally
We added Turbo and left it on globally which is the default.

This introduced a bug with the parental consent journey that causes
users to not reach the confirmation page, which might lead them to think
that the service is broken.

Turbo allows an alternate mode of operation where it's off by default
and can be introduced on a per-link basis with the `data-turbo="true"`
attribute: https://turbo.hotwired.dev/handbook/drive#disabling-turbo-drive-on-specific-links-or-forms