Remove JSON example campaign code and data
We've created a new `/reset` endpoint that generates an example
campaign, and have removed the Playwright tests that relied on these
JSON example campaigns. Removing these does mean that if we need to
generate static/predictable campaign data again in the future, we'll
need to develop a new way of doing it, but we'd prefer to do this with
the factories directly instead of using an intermediary JSON file.