Ensure patient_relationship exists on factory'd consents
This is a quick tidy-up of the factories with the main aim of fixing an
issue with our generated data that caused consents not to have any
`parent_relationship` defined.

To this end, `consents` now always inherit the parent information from
the `patient` unless otherwise specified via `:from_mum` or `:from_dad`.

A few other changes:

- Some redundant traits were removed from the `patient` factory that
  created transient `patient_sessions`. The call-sites of these have
  been replaced with calls to the `patient_session` factory instead
- All patients now have an `address` by default as that's the more
  common case
- All patients now have a default age between 12-13 as that's the more
  common case