Add rufo to format erb files
We already use `prettier` for Ruby, JS/TS, (S)CSS, plain HTML, markdown,
and yml, but unfortunately it doesn't support `.erb` which is a
significant portion of our codebase.

[`rufo`](https://github.com/ruby-formatter/rufo) does however and it has
good editor support.

It's not as deterministic as `prettier` and there are some edge cases
where it "changes its mind" and formats identical input in two different
ways. But for the most part it seems to work well enough.