Add simple PDS patient model
This is the simplest possible way of getting a patient from PDS. Wrapped
up into a model so we can change the implementation later.

The model itself is also quite simple, whereas the FHIR model may
potentially have multiple names (in case the patient has changed their
name), etc. We need to better understand how Mavis uses this data before
we expand our own model, so consider this a proof-of-concept.