Add "or" divider to triage form options
Also update the legend styling to match the prototype. It should be
bolded on the patient page, and on the update triage outcome page it
should not be. For the latter, we need some custom CSS as the
`govuk_design_system_formbuilder` one always inserts a fieldset size
which sets the font weight to bold and adds a margin.