Rename sign in and sign out
In the designs these have been renamed to "Log in" and "Log out". I've
decided internally to keep the naming "Sign X" as that matches our
database column names and the naming convention that Devise uses.

Although I've changed the content on the log in page as well, this
design is likely to change anyway as we integrate with CIS2.