Use triage form component in verbal consent flow
We were implementing this form separately and duplicating the radio
buttons before. Reusing the component requires making some of the
options more flexible. This adds the ability to specify a hidden legend,
as well as overriding the form method. The content of the submit button
is also now conditional on the form method.