Enable frozen string literals
This modifies the Rubocop rules to enforce a frozen string literal
comment at the top of every Ruby file, and then runs Rubocop with
autocorrect to add one to each file.

Frozen string literals was going to be default in Ruby 3.0, but this
didn't happen in the end due to concerns about incompatibility, however
I think this is good practice for new projects. An alternative to the
comment is using the `--enable=frozen-string-literal` flag, however this
applies to third party Gems and the standard library, which may then
introduce the compatibility issues.

Aside from not needing to manually call `.freeze` on string literals,
this approach introduces a small performance increase (although unlikely
to be particularly noticable). However, it does reduce the chance of
bugs happening related to mutable string literals.