Ignore scratchpad dir in Rubocop
This is a convention we've adopted for a local directory to store
project-related files that we don't want to add to the repo. It's
ignored in the other files, but recently I've added some dirty, smelly
Ruby code that even Rubocop couldn't fix for me, and its errors were
blocking running bin/lint, so let's just ignore it.