Use traits_for_enum in triage factory
This updates the triage factory to use the `traits_for_enum` helper, and
renames the factory traits to ensure they match the status names. This
makes the code clearer by using a consistent naming, if we want to
rename one of the statuses then we can do that on the model itself.