Rename nhsuk_date_day_of_week
This has been renamed to `long_day_of_week` as it's very similar to the
standard `long` date format, but with the date of the week included. By
re-using the standard date format names we ensure that any third party
Gems use the same formats as we do across the service.