Rename nhs_date date formatter
This is very similar to the built in `long` format, except with the
full day of the week, so instead of introducing our own format I suggest
we replace the built in `long` format. By re-using the standard date
format names we ensure that any third party Gems use the same formats
as we do across the service.