MAV-1212 Fail confirmation when child is not attending.
Ideally we would get the error message to be a direct link to the
child on the registration page, see
https://github.com/x-govuk/govuk-form-builder/issues/566
for why this is not yet possible.

I could not find a way to test this feature with an e2e test as our
testing library uses a mocked browser that does not actually allow you
to open a second tab in which to modify the attendance while submitting
the vaccination.