Add vaccine available_delivery_sites/_methods
To accomodate NIVS upload, we need to relax the enums by allowing
left/right buttocks for delivery sites, as well as a nasal spray method.

These are not currently in use by any of our vaccines. The simplest
solution is to create a couple of new methods that remove the unused
options and can be used when recording a vaccination to present the
available sites and methods.

The logic in these methods will need to become more complex at some
point to accommodate different types of vaccines.