Allow passing text to component
This updates the `AppSecondaryNavigationComponent` to support passing in
a `text` argument as an alternative to passing in a block. This is just
a convenience, but I think it makes it easier to read when this
component is called. It's a common pattern that already exists in the
GOV.UK components Gem.