Add gender code to patients
This will be used to capture the "person stated gender code" for
patients, coming from the immunisation imports.

Ideally we wouldn't need to capture gender, and if we did it's
recommended instead to store a gender identity code, but the data we're
dealing with uses the gender code. I did consider if we could map the
gender code values to the gender identity code values, but it's not a
complete one to one mapping (specifically it's unclear how to map the
non specified or indeterminate gender code to a gender identity).

https://digital.nhs.uk/data-and-information/data-collections-and-data-sets/data-sets/mental-health-services-data-set/submit-data/data-quality-of-protected-characteristics-and-other-vulnerable-groups/gender-identity#how-to-collect-and-record-gender-identity-data

https://service-manual.nhs.uk/content/inclusive-content/sex-gender-and-sexuality