Only export unexported vaccination_records
The DPS export feature will track what records have been exported and
only export those that haven't already been exported. There will be
logic in the future to allow for updates, so for now it seems to make
sense to use a timestamp to track if and when the export was done.