Make batch expiry and name not nullable
These fields should always be populated as we've got validation on the
model saying they must be present, so we can safely mark them as not
nullable, reducing the risk of a null value being stored by mistake.

Unrelated, this also updates the schema comments for the patient model
as an enum was added after the comments were written.