Include FactoryBot methods in previews
This fixes an issue that was introduced in
3869396d1d696cea865e2be15e7da5a3bfe9e6a6 where the components could no
longer be rendered as they were referring to FactoryBot methods
implicitly which are not automatically included in previews.

I was hoping there was a way of including this automatically in all
previews, but aside from creating our own ViewComponent::Preview class
there doesn't seem to be a nice way of doing it.