Move NIVS report to vaccination records controller
This refactors the code related to the existing NIVS report following
the introduction of the vaccination records tab on the campaigns. By
moving the report generation code in to the vaccination records
controller, this allows us to re-use some of the existing logic around
fetching vaccination records to ensure that the output matches the
records that are shown to the user on the tab.