Handle variations in gender code column name
We expect that some import CSV will use the column `GENDER_CODE` and
some will use `GENDER` so we need to update the importer to handle both.

I've also removed this column from the required columns validation as it
could be either. This will still raise a row-level validation error on
the gender if neither column exists, but we might want to improve the
header validation in the future to handle optional values.