Add vaccines initializer
This is a translation to `.yml` of the vaccine configuration table in
the Mavis reporting spec.

Immunisation import rows contain a reference to a vaccine name which can
roughly be constructed or linked to one of these records.

By adding this to the application, we can use it to:

- Validate immunisation import rows and map them to the canonical
  product codes
- Present a list of vaccine options to choose from when creating a
  campaign
- Improve our factories and seed data with more realistic vaccines

The vaccines are keyed by product_code which is also reflected in the
values of each, which could come in handy.