Import dose sequence from HPV files
For HPV, it's valid to have a dose sequence and therefore we need to
ensure that it's imported correctly. This is an optional field, as Flu
files won't contain a dose sequence.

To determine whether we need to import the dose sequence or not we use
the maximum dose sequence on the vaccine.