Refactor how records are imported
Rather than importing all the patients, all the locations, all the
sessions one by one, we instead import the entire row by one, each
location, patient, session, etc one at a time.

This allows us to import files which might make reference to multiple
patients.

Although the current approach (of applying a unique filter on the models
before importing them) works, it no long applies correctly for patients
as there isn't a single field we can use to de-duplicate them. This used
to be the NHS number, but now we allow the NHS number to be `nil`.

The current approach also ends up firing multiple calls to the database
to look up locations which were previously just created, so this should
provide a small performance increase on the implementation.