Add disable_ssl_verification? setting
This is used when connecting to a locally running sandbox instance of
MESH. The reason this is in settings is so that it is kept close to the
host setting, since this should only be enabled when that is
`localhost`. A feature flag was considered and discarded, since it
doesn't make sense for this to be changed separately from the MESH host
setting.