Don't import un-administered records
When importing data for the immunisation import we want to ignore
records where the vaccine wasn't administered, rather than importing
them. This is to match the Mavis reporting specification.

For now we should another flash banner with the number of ignored
records, but we might want to improve the UI in the future.