Rename patient location to school
This is to remove any ambiguity over what the patient location refers
to, it should represent the school that the patient belongs to. I
haven't renamed the model itself since various non-school locations can
be stored in this table.

I've also introduced a helper method for rendering the name of the
patient's school (or "unknown school" if it's nil). This will be
expanded in the future to support when a patient is home schooled.