Don't assume HPV in campaign factory
This changes the campaign test factory to not assume an HPV campaign by
default and instead create a simpler campaign with fewer associations.
Locally this improves the performance of the tests by about 6 seconds
(from 71 seconds to 65 seconds), and it also helps us to write tests
that don't assume a particular type of campaign.