Ensure campaign name is unique
Within the same academic year, type and team. This avoids a situation
where users create campaigns for the same year, vaccine and team and
with the same name, making them difficult to distinguish within the
interface.

I've also normalised the name to ensure it doesn't contain whitespace on
either side of the string.