Show vaccination administration date
This updates the vaccination details component to show the date of when
the vaccination was administered. We're not showing the time component
because we don't have one in the import so it's always set to midnight.

In the future we could amend this component to show the time component
if we have one (i.e. the vaccination was recorded in the service
directly).