Remove unneeded presence validations
When a presence validation is paired with an inclusion validation
there's often no need to have both. The inclusion implicitly checks
whether the value is present by requiring it to match the list of valid
values, and we often don't need to present the user with two error
messages (one for presence and one for inclusion).