Process imported school URLs
The GIAS CSV contains some school addresses that don't have a protocol.
As such, they are not valid URLs according to validate_url.

To work around this, we can force prepend `https://`, which assumes that
the school websites have HTTPS setup. This isn't a given, but it feels
like the better default, since HTTPS is pretty strongly pushed by
browser vendors.

The URL for Legh Vale Primary School is also plainly malformed and
starts with `http:www`. It's the only instance of this, but we have to
work around it.