Handle session location being nil
This is now a valid state for sessions that have been imported as some
of the rows won't match with a known location. I've updated most cases
where we show the session location to use a new helper that handles the
value being `nil`.

I haven't updated all places as there are many parts of the service that
require the location to be present, and for records created in the
service this will always be true, it's only those that were imported
where it's possible.