Store performed by name on vaccination records
This is an optional set of fields for storing the given name and family
name of the person who performed a vaccination. This is required for the
import as we may be importing records for users that don't exist in the
service, and with just a given/family name we can't match with our users
as the information isn't unique enough (two different people could share
the same name).