Remove default scopes
A number of models use default scopes to automatically only show
"active" or "recorded" instances. This can be useful, but as the
codebase grows it can start to cause unexpected behaviour and the need
to override the default scope becomes more common.

I've attempted to find all the places where the default scopes were
being used and update them in this commit, but it's possible some of
them may have been missed and will need to be updated in the future.