Add "check and confirm" import page
This adds an additional page when importing records that allows the user
to review the records before confirming the import.

I've also amended how we import vaccination records and sessions to
ensure that we don't ignore duplicate records if the file had already
been read in but not confirmed, whilst also preventing the same file
from importing the same records multiple times.