Show immunisation import warning callout
If there are duplicate records or not administered records, we want to
include this information in a warning callout at the top of the page.

The warning callout component is part of the NHS design system, but not
part of GOV.UK's so we don't have a component in govuk-components for
it. As this is the first time we're using it, I decided against creating
a component for it, but we might want to do add one in the future.