Replace form_for with form_with
This replaces all uses of the form_for form tag helper with form_with
which was introduced in Rails 5.1 and is generally recommended as the
approach going forward. Although the current approach works and hasn't
been fully deprecated, I think it makes sense to use the newer approach.

See https://guides.rubyonrails.org/form_helpers.html#using-form-tag-and-form-for
for more details.