Ensure form errors appear at the top of the page
This refactors how we render the form error summary component to ensure
that it always appears at the top of the page, regardless of the
structure of the form.

This means either using a `content_for(:before_content)` block or moving
any tags shown above the form to below the call to
`f.govuk_error_summary`.