Add ability to edit vaccination record date/time
This adds the ability to edit the date and time an existing vaccination
record.

I've decided to add these actions to a new controller as it's likely
we'll need to edit different parts of the vaccination records in the
future. I've also set up the controller in a similar way to how we
handle wizards as I imagine we'll likely need that functionality in the
future.

The govuk_form_builder Gem doesn't currently support time fields (and
there's no mention of how they should work in the design system) so I've
had to re-create it myself in the form.

If the user is on this page we should expect the date/time to always be
set, whereas we allow the field to be optional on the model. One way to
solve this could be to use a form object, however for this page I've
decided to add some validation logic to the controller. We could
refactor this in to a form object in the future.