Rename form_steps to wizard_steps
We're currently using the term `form_steps` in a number of models to
refer to the various steps of wizards in the service. This is used in
the Wicked controllers, and in the `on_wizard_step` validation helper.

This commit renames it to `wizard_steps` to be consistent in our naming
approach towards wizards, sticking with only one name.