Add changing campaign information
This adds the ability to change the existing information on a campaign,
using the same controller that we have for creating new campaigns. This
is built by adding "Change" links to the summary list for the campaign,
and ensuring we skip straight to the confirmation step rather than the
next step.