Add changing campaign vaccines
This adds a new step for editing campaigns where the user can choose the
vaccines that belong to a campaign. When creating a campaign the
vaccines default to the active vaccines for a particular programme type,
but we may need to add discontinued vaccines to campaigns that happened
in the past, to support importing historical records.