Import draft patient sessions
When first importing immunisations we want the patient sessions to be
draft so they don't appear in any parts of the interface, until the
import is confirmed and then they are marked as active.

Without this, it's possible for a patient to appear in parts of the
system if the session already existed (and was active).