Refactor PatientSession and associated factories
This refactors the factories related to patient sessions, and therefore
sessions and patients too, to simplify their use and improve the code
readability.

The main changes are automatic creation of patient sessions if a session
is passed to the patient when created, which simplifies a number of
other factories; and the use of association in the factories.